Adding feature to check if system is unix or windows
Once this check passes, it will use commands specific command to it\'s operating shell. For example, at the command `rm -rf %s` would not work on either terminals, cause -rf is not a valid argument, the command exists for powershell, but not on the normal windows command line. So you would have to change that to `DEL` command isntead, which is supported by both powershell and windows command line.\n\nAnother fix is added is mkdir, though mkdir exists for the commandline, it returns an error when the folder already exists, and the -p argument is not valid for mkdir, and instead creates an actual folder called `-p`. This is why we use `if not exist %s mkdir %s`, cause it checks if that folder exists before creating it. This works on the windows command line, but not on powershell, but if we can, we could also add a check for powershell and use a powershell command instead.